requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
pandas==1.3.3
openpyxl==3.0.9
bcrypt==3.2.0
requests==2.26.0
python-dotenv==0.19.0
cloudinary==1.33.0
